* os.unsetenv() may be present but a fake (e.g. there is no nt.unsetenv()).   Fix the detection code.
* rbuiltin.py was not tracking updates to the extfunctables, causing errors
  if you ran in the same process an rpython test followed by
  translator/c/test/test_extfunc.

+++ pypy/rpython/module/test/test_ll_os.py
@@ -46,14 +46,16 @@ def test_putenv_unsetenv():
     assert result == '12345678'
     f.close()
     os.unlink(filename)
-    ll_os_unsetenv(to_rstr("abcdefgh"))
-    cmd = '''python -c "import os; print repr(os.getenv('abcdefgh'))" > %s''' % filename
-    os.system(cmd)
-    f = file(filename)
-    result = f.read().strip()
-    assert result == 'None'
-    f.close()
-    os.unlink(filename)
+    posix = __import__(os.name)
+    if hasattr(posix, "unsetenv"):
+        ll_os_unsetenv(to_rstr("abcdefgh"))
+        cmd = '''python -c "import os; print repr(os.getenv('abcdefgh'))" > %s''' % filename
+        os.system(cmd)
+        f = file(filename)
+        result = f.read().strip()
+        assert result == 'None'
+        f.close()
+        os.unlink(filename)
 
 test_src = """
 import os
